storage: more assorted code movement
Release note: None
