+++ pkg/storage/replica.go
@@ -70,7 +70,7 @@ import (
 
 const (
 	// configGossipTTL is the time-to-live for configuration maps.
-	configGossipTTL = 0 // does not expire
+
 	// optimizePutThreshold is the minimum length of a contiguous run
 	// of batched puts or conditional puts, after which the constituent
 	// put operations will possibly be optimized by determining whether
@@ -3725,204 +3725,6 @@ func evaluateBatch(
 	return br, result, nil
 }
 
-// getLeaseForGossip tries to obtain a range lease. Only one of the replicas
-// should gossip; the bool returned indicates whether it's us.
-func (r *Replica) getLeaseForGossip(ctx context.Context) (bool, *roachpb.Error) {
-	// If no Gossip available (some tests) or range too fresh, noop.
-	if r.store.Gossip() == nil || !r.IsInitialized() {
-		return false, roachpb.NewErrorf("no gossip or range not initialized")
-	}
-	var hasLease bool
-	var pErr *roachpb.Error
-	if err := r.store.Stopper().RunTask(
-		ctx, "storage.Replica: acquiring lease to gossip",
-		func(ctx context.Context) {
-			// Check for or obtain the lease, if none active.
-			_, pErr = r.redirectOnOrAcquireLease(ctx)
-			hasLease = pErr == nil
-			if pErr != nil {
-				switch e := pErr.GetDetail().(type) {
-				case *roachpb.NotLeaseHolderError:
-					// NotLeaseHolderError means there is an active lease, but only if
-					// the lease holder is set; otherwise, it's likely a timeout.
-					if e.LeaseHolder != nil {
-						pErr = nil
-					}
-				default:
-					// Any other error is worth being logged visibly.
-					log.Warningf(ctx, "could not acquire lease for range gossip: %s", e)
-				}
-			}
-		}); err != nil {
-		pErr = roachpb.NewError(err)
-	}
-	return hasLease, pErr
-}
-
-// maybeGossipFirstRange adds the sentinel and first range metadata to gossip
-// if this is the first range and a range lease can be obtained. The Store
-// calls this periodically on first range replicas.
-func (r *Replica) maybeGossipFirstRange(ctx context.Context) *roachpb.Error {
-	if !r.IsFirstRange() {
-		return nil
-	}
-
-	// When multiple nodes are initialized with overlapping Gossip addresses, they all
-	// will attempt to gossip their cluster ID. This is a fairly obvious misconfiguration,
-	// so we error out below.
-	if gossipClusterID, err := r.store.Gossip().GetClusterID(); err == nil {
-		if gossipClusterID != r.store.ClusterID() {
-			log.Fatalf(
-				ctx, "store %d belongs to cluster %s, but attempted to join cluster %s via gossip",
-				r.store.StoreID(), r.store.ClusterID(), gossipClusterID)
-		}
-	}
-
-	// Gossip the cluster ID from all replicas of the first range; there
-	// is no expiration on the cluster ID.
-	if log.V(1) {
-		log.Infof(ctx, "gossiping cluster id %q from store %d, r%d", r.store.ClusterID(),
-			r.store.StoreID(), r.RangeID)
-	}
-	if err := r.store.Gossip().AddClusterID(r.store.ClusterID()); err != nil {
-		log.Errorf(ctx, "failed to gossip cluster ID: %s", err)
-	}
-
-	if r.store.cfg.TestingKnobs.DisablePeriodicGossips {
-		return nil
-	}
-
-	hasLease, pErr := r.getLeaseForGossip(ctx)
-	if pErr != nil {
-		return pErr
-	} else if !hasLease {
-		return nil
-	}
-	r.gossipFirstRange(ctx)
-	return nil
-}
-
-// maybeSetCorrupt is a stand-in for proper handling of failing replicas. Such a
-// failure is indicated by a call to maybeSetCorrupt with a ReplicaCorruptionError.
-// Currently any error is passed through, but prospectively it should stop the
-// range from participating in progress, trigger a rebalance operation and
-// decide on an error-by-error basis whether the corruption is limited to the
-// range, store, node or cluster with corresponding actions taken.
-//
-// TODO(d4l3k): when marking a Replica corrupt, must subtract its stats from
-// r.store.metrics. Errors which happen between committing a batch and sending
-// a stats delta from the store are going to be particularly tricky and the
-// best bet is to not have any of those.
-// @bdarnell remarks: Corruption errors should be rare so we may want the store
-// to just recompute its stats in the background when one occurs.
-func (r *Replica) maybeSetCorrupt(ctx context.Context, pErr *roachpb.Error) *roachpb.Error {
-	if cErr, ok := pErr.GetDetail().(*roachpb.ReplicaCorruptionError); ok {
-		r.mu.Lock()
-		defer r.mu.Unlock()
-
-		log.Errorf(ctx, "stalling replica due to: %s", cErr.ErrorMsg)
-		cErr.Processed = true
-		r.mu.destroyStatus.Set(cErr, destroyReasonCorrupted)
-		pErr = roachpb.NewError(cErr)
-
-		// Try to persist the destroyed error message. If the underlying store is
-		// corrupted the error won't be processed and a panic will occur.
-		if err := r.mu.stateLoader.SetReplicaDestroyedError(ctx, r.store.Engine(), pErr); err != nil {
-			cErr.Processed = false
-			return roachpb.NewError(cErr)
-		}
-
-		log.Fatalf(ctx, "replica is corrupted: %s", pErr)
-	}
-	return pErr
-}
-
-// SplitByLoadEnabled wraps "kv.range_split.by_load_enabled".
-var SplitByLoadEnabled = settings.RegisterBoolSetting(
-	"kv.range_split.by_load_enabled",
-	"allow automatic splits of ranges based on where load is concentrated.",
-	true,
-)
-
-// SplitByLoadQPSThreshold wraps "kv.range_split.load_qps_threshold".
-var SplitByLoadQPSThreshold = settings.RegisterIntSetting(
-	"kv.range_split.load_qps_threshold",
-	"the QPS over which, the range becomes a candidate for load based splitting.",
-	250, // 250 req/s
-)
-
-// SplitByLoadQPSThreshold returns the QPS request rate for a given replica.
-func (r *Replica) SplitByLoadQPSThreshold() float64 {
-	return float64(SplitByLoadQPSThreshold.Get(&r.store.cfg.Settings.SV))
-}
-
-// SplitByLoadEnabled returns whether load based splitting is enabled.
-func (r *Replica) SplitByLoadEnabled() bool {
-	return SplitByLoadEnabled.Get(&r.store.cfg.Settings.SV) &&
-		r.store.ClusterSettings().Version.IsMinSupported(cluster.VersionLoadSplits) &&
-		!r.store.TestingKnobs().DisableLoadBasedSplitting
-}
-
-// needsSplitByLoadLocked returns two bools indicating first, whether
-// the range is over the threshold for splitting by load and second,
-// whether the range is ready to be added to the split queue.
-func (r *Replica) needsSplitByLoadLocked() (bool, bool) {
-	// First compute requests per second since the last check.
-	nowTime := r.store.Clock().PhysicalTime()
-	duration := nowTime.Sub(r.splitMu.lastReqTime)
-	if duration < time.Second {
-		return r.splitMu.splitFinder != nil, false
-	}
-
-	// Update the QPS and reset the time and request counter.
-	r.splitMu.qps = (float64(r.splitMu.count) / float64(duration)) * 1e9
-	r.splitMu.lastReqTime = nowTime
-	r.splitMu.count = 0
-
-	// If the QPS for the range exceeds the threshold, start actively
-	// tracking potential for splitting this range based on load.
-	// This tracking will begin by initiating a splitFinder so it can
-	// begin to Record requests so it can find a split point. If a
-	// splitFinder already exists, we check if a split point is ready
-	// to be used.
-	if r.splitMu.qps >= r.SplitByLoadQPSThreshold() {
-		if r.splitMu.splitFinder != nil {
-			if r.splitMu.splitFinder.Ready(nowTime) {
-				// We're ready to add this range to the split queue.
-				return true, true
-			}
-		} else {
-			r.splitMu.splitFinder = split.New(nowTime)
-		}
-		return true, false
-	}
-
-	r.splitMu.splitFinder = nil
-	return false, false
-}
-
-// needsSplitBySize returns true if the size of the range requires it
-// to be split.
-func (r *Replica) needsSplitBySize() bool {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	return r.needsSplitBySizeRLocked()
-}
-
-func (r *Replica) needsSplitBySizeRLocked() bool {
-	return r.exceedsMultipleOfSplitSizeRLocked(1)
-}
-
-func (r *Replica) needsMergeBySizeRLocked() bool {
-	return r.mu.state.Stats.Total() < *r.mu.zone.RangeMinBytes
-}
-
-func (r *Replica) exceedsMultipleOfSplitSizeRLocked(mult float64) bool {
-	maxBytes := *r.mu.zone.RangeMaxBytes
-	size := r.mu.state.Stats.Total()
-	return maxBytes > 0 && float64(size) > float64(maxBytes)*mult
-}
-
 func (r *Replica) startKey() roachpb.RKey {
 	return r.Desc().StartKey
 }
@@ -3930,15 +3732,6 @@ func (r *Replica) startKey() roachpb.RKey {
 // Less implements the btree.Item interface.
 func (r *Replica) Less(i btree.Item) bool {
 	return r.startKey().Less(i.(rangeKeyItem).startKey())
-}
-
-func isRaftLeader(raftStatus *raft.Status) bool {
-	return raftStatus != nil && raftStatus.SoftState.RaftState == raft.StateLeader
-}
-
-// HasRaftLeader returns true if the raft group has a raft leader currently.
-func HasRaftLeader(raftStatus *raft.Status) bool {
-	return raftStatus != nil && raftStatus.SoftState.Lead != 0
 }
 
 // GetLeaseHistory returns the lease history stored on this replica.
