+++ pkg/storage/replica_raft.go
@@ -2153,3 +2153,12 @@ func (r *Replica) getAndGCSnapshotLogTruncationConstraintsLocked(
 	}
 	return minSnapIndex
 }
+
+func isRaftLeader(raftStatus *raft.Status) bool {
+	return raftStatus != nil && raftStatus.SoftState.RaftState == raft.StateLeader
+}
+
+// HasRaftLeader returns true if the raft group has a raft leader currently.
+func HasRaftLeader(raftStatus *raft.Status) bool {
+	return raftStatus != nil && raftStatus.SoftState.Lead != 0
+}
